+++ simtools/__init__.py
@@ -7,3 +7,6 @@ Python.
 
 __version__ = '0.1.0.dev1'
 __author__ = "Przemyslaw (Mack) Nowak"
+
+from .params import load_params, Params
+from . import params