+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Added loading parameters from a Python file.
+- Added hierarchy of SimTools exceptions.
 - Added package installer.
 - Added package description.
 - Added package license terms.